services/catalog/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecom-catalog"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "pydantic==2.11.7",
    "sqlalchemy==2.0.43",
    "alembic==1.16.4",
    "psycopg[binary]==3.2.9",
    "python-multipart==0.0.20",
    "minio==7.2.16",
    "PyJWT==2.10.1",
    "prometheus-fastapi-instrumentator==7.1.0",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "httpx", "requests"]
dev = ["pytest", "black", "flake8", "mypy", "pytest-asyncio"]

[tool.setuptools.packages.find]
where = [""]
include = ["app*"]